+++ src/voku/SimplePhpParser/Parsers/PhpCodeParser.php
@@ -54,6 +54,20 @@ final class PhpCodeParser
         array $autoloaderProjectPaths = [],
         array $pathExcludeRegex = []
     ): ParserContainer {
+        foreach ($autoloaderProjectPaths as $projectPath) {
+            if (\file_exists($projectPath . '/vendor/autoload.php')) {
+                /** @noinspection PhpIncludeInspection */
+                require_once $projectPath . '/vendor/autoload.php';
+            } elseif (\file_exists($projectPath . '/../vendor/autoload.php')) {
+                /** @noinspection PhpIncludeInspection */
+                require_once $projectPath . '/../vendor/autoload.php';
+            } elseif (\file_exists($projectPath) && \is_file($projectPath)) {
+                /** @noinspection PhpIncludeInspection */
+                require_once $projectPath;
+            }
+        }
+        restore_error_handler();
+
         $phpCodes = self::getCode(
             $pathOrCode,
             $pathExcludeRegex
@@ -183,6 +197,7 @@ final class PhpCodeParser
                 require_once $projectPath;
             }
         }
+        restore_error_handler();
 
         new \voku\SimplePhpParser\Parsers\Helper\Psalm\FakeFileProvider();
         $providers = new \Psalm\Internal\Provider\Providers(